+++ Resources/roguelike-16x16.tsx
@@ -498,4 +498,22 @@
    <property name="type" value="water"/>
   </properties>
  </tile>
+ <tile id="1767">
+  <properties>
+   <property name="walkable" type="bool" value="true"/>
+   <property name="weight" type="float" value="-5"/>
+  </properties>
+ </tile>
+ <tile id="1768">
+  <properties>
+   <property name="walkable" type="bool" value="true"/>
+   <property name="weight" type="float" value="1"/>
+  </properties>
+ </tile>
+ <tile id="1769">
+  <properties>
+   <property name="walkable" type="bool" value="true"/>
+   <property name="weight" type="float" value="200"/>
+  </properties>
+ </tile>
 </tileset>